requirements.txt
customtkinter
openai-whisper
deep-translator
pillow